LayoutTests/fast/selectors/nth-child-of-selector-list-expected.txt

Check ":nth-child(An+B of selectorList)" with a selector list.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing ":nth-child(2n of *, :not(*), *)"
PASS document.querySelectorAll('#test-root :nth-child(2n of *, :not(*), *)').length is 15
PASS document.querySelectorAll('#test-root :nth-child(2n of *, :not(*), *)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :nth-child(2n of *, :not(*), *)')[1].id is "testcase4"
PASS document.querySelectorAll('#test-root :nth-child(2n of *, :not(*), *)')[2].id is "testcase6"
PASS document.querySelectorAll('#test-root :nth-child(2n of *, :not(*), *)')[3].id is "testcase8"
PASS document.querySelectorAll('#test-root :nth-child(2n of *, :not(*), *)')[4].id is "testcase10"
PASS document.querySelectorAll('#test-root :nth-child(2n of *, :not(*), *)')[5].id is "testcase12"
PASS document.querySelectorAll('#test-root :nth-child(2n of *, :not(*), *)')[6].id is "testcase14"
PASS document.querySelectorAll('#test-root :nth-child(2n of *, :not(*), *)')[7].id is "testcase16"
PASS document.querySelectorAll('#test-root :nth-child(2n of *, :not(*), *)')[8].id is "testcase18"
PASS document.querySelectorAll('#test-root :nth-child(2n of *, :not(*), *)')[9].id is "testcase20"
PASS document.querySelectorAll('#test-root :nth-child(2n of *, :not(*), *)')[10].id is "testcase22"
PASS document.querySelectorAll('#test-root :nth-child(2n of *, :not(*), *)')[11].id is "testcase24"
PASS document.querySelectorAll('#test-root :nth-child(2n of *, :not(*), *)')[12].id is "testcase26"
PASS document.querySelectorAll('#test-root :nth-child(2n of *, :not(*), *)')[13].id is "testcase28"
PASS document.querySelectorAll('#test-root :nth-child(2n of *, :not(*), *)')[14].id is "testcase30"
PASS styledElements is ["testcase2","testcase4","testcase6","testcase8","testcase10","testcase12","testcase14","testcase16","testcase18","testcase20","testcase22","testcase24","testcase26","testcase28","testcase30"]

Testing ":nth-child(2n of :not(*), :empty:not(:empty), [attribute1][attribute2][attribute3], .foo.bar.baz.bazoo, :not([id]), :not([class]))"
PASS document.querySelectorAll('#test-root :nth-child(2n of :not(*), :empty:not(:empty), [attribute1][attribute2][attribute3], .foo.bar.baz.bazoo, :not([id]), :not([class]))').length is 0
PASS styledElements is []

Testing ":nth-child(odd of testcaseA, testcaseB)"
PASS document.querySelectorAll('#test-root :nth-child(odd of testcaseA, testcaseB)').length is 15
PASS document.querySelectorAll('#test-root :nth-child(odd of testcaseA, testcaseB)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :nth-child(odd of testcaseA, testcaseB)')[1].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(odd of testcaseA, testcaseB)')[2].id is "testcase5"
PASS document.querySelectorAll('#test-root :nth-child(odd of testcaseA, testcaseB)')[3].id is "testcase7"
PASS document.querySelectorAll('#test-root :nth-child(odd of testcaseA, testcaseB)')[4].id is "testcase9"
PASS document.querySelectorAll('#test-root :nth-child(odd of testcaseA, testcaseB)')[5].id is "testcase11"
PASS document.querySelectorAll('#test-root :nth-child(odd of testcaseA, testcaseB)')[6].id is "testcase13"
PASS document.querySelectorAll('#test-root :nth-child(odd of testcaseA, testcaseB)')[7].id is "testcase15"
PASS document.querySelectorAll('#test-root :nth-child(odd of testcaseA, testcaseB)')[8].id is "testcase17"
PASS document.querySelectorAll('#test-root :nth-child(odd of testcaseA, testcaseB)')[9].id is "testcase19"
PASS document.querySelectorAll('#test-root :nth-child(odd of testcaseA, testcaseB)')[10].id is "testcase21"
PASS document.querySelectorAll('#test-root :nth-child(odd of testcaseA, testcaseB)')[11].id is "testcase23"
PASS document.querySelectorAll('#test-root :nth-child(odd of testcaseA, testcaseB)')[12].id is "testcase25"
PASS document.querySelectorAll('#test-root :nth-child(odd of testcaseA, testcaseB)')[13].id is "testcase27"
PASS document.querySelectorAll('#test-root :nth-child(odd of testcaseA, testcaseB)')[14].id is "testcase29"
PASS styledElements is ["testcase1","testcase3","testcase5","testcase7","testcase9","testcase11","testcase13","testcase15","testcase17","testcase19","testcase21","testcase23","testcase25","testcase27","testcase29"]

Testing ":nth-child(even of .foo, .bar, .baz)"
PASS document.querySelectorAll('#test-root :nth-child(even of .foo, .bar, .baz)').length is 15
PASS document.querySelectorAll('#test-root :nth-child(even of .foo, .bar, .baz)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :nth-child(even of .foo, .bar, .baz)')[1].id is "testcase4"
PASS document.querySelectorAll('#test-root :nth-child(even of .foo, .bar, .baz)')[2].id is "testcase6"
PASS document.querySelectorAll('#test-root :nth-child(even of .foo, .bar, .baz)')[3].id is "testcase8"
PASS document.querySelectorAll('#test-root :nth-child(even of .foo, .bar, .baz)')[4].id is "testcase10"
PASS document.querySelectorAll('#test-root :nth-child(even of .foo, .bar, .baz)')[5].id is "testcase12"
PASS document.querySelectorAll('#test-root :nth-child(even of .foo, .bar, .baz)')[6].id is "testcase14"
PASS document.querySelectorAll('#test-root :nth-child(even of .foo, .bar, .baz)')[7].id is "testcase16"
PASS document.querySelectorAll('#test-root :nth-child(even of .foo, .bar, .baz)')[8].id is "testcase18"
PASS document.querySelectorAll('#test-root :nth-child(even of .foo, .bar, .baz)')[9].id is "testcase20"
PASS document.querySelectorAll('#test-root :nth-child(even of .foo, .bar, .baz)')[10].id is "testcase22"
PASS document.querySelectorAll('#test-root :nth-child(even of .foo, .bar, .baz)')[11].id is "testcase24"
PASS document.querySelectorAll('#test-root :nth-child(even of .foo, .bar, .baz)')[12].id is "testcase26"
PASS document.querySelectorAll('#test-root :nth-child(even of .foo, .bar, .baz)')[13].id is "testcase28"
PASS document.querySelectorAll('#test-root :nth-child(even of .foo, .bar, .baz)')[14].id is "testcase30"
PASS styledElements is ["testcase2","testcase4","testcase6","testcase8","testcase10","testcase12","testcase14","testcase16","testcase18","testcase20","testcase22","testcase24","testcase26","testcase28","testcase30"]

Testing ":nth-child(odd of [attribute1], [attribute2=value2])"
PASS document.querySelectorAll('#test-root :nth-child(odd of [attribute1], [attribute2=value2])').length is 15
PASS document.querySelectorAll('#test-root :nth-child(odd of [attribute1], [attribute2=value2])')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root :nth-child(odd of [attribute1], [attribute2=value2])')[1].id is "testcase3"
PASS document.querySelectorAll('#test-root :nth-child(odd of [attribute1], [attribute2=value2])')[2].id is "testcase5"
PASS document.querySelectorAll('#test-root :nth-child(odd of [attribute1], [attribute2=value2])')[3].id is "testcase7"
PASS document.querySelectorAll('#test-root :nth-child(odd of [attribute1], [attribute2=value2])')[4].id is "testcase9"
PASS document.querySelectorAll('#test-root :nth-child(odd of [attribute1], [attribute2=value2])')[5].id is "testcase11"
PASS document.querySelectorAll('#test-root :nth-child(odd of [attribute1], [attribute2=value2])')[6].id is "testcase13"
PASS document.querySelectorAll('#test-root :nth-child(odd of [attribute1], [attribute2=value2])')[7].id is "testcase15"
PASS document.querySelectorAll('#test-root :nth-child(odd of [attribute1], [attribute2=value2])')[8].id is "testcase17"
PASS document.querySelectorAll('#test-root :nth-child(odd of [attribute1], [attribute2=value2])')[9].id is "testcase19"
PASS document.querySelectorAll('#test-root :nth-child(odd of [attribute1], [attribute2=value2])')[10].id is "testcase21"
PASS document.querySelectorAll('#test-root :nth-child(odd of [attribute1], [attribute2=value2])')[11].id is "testcase23"
PASS document.querySelectorAll('#test-root :nth-child(odd of [attribute1], [attribute2=value2])')[12].id is "testcase25"
PASS document.querySelectorAll('#test-root :nth-child(odd of [attribute1], [attribute2=value2])')[13].id is "testcase27"
PASS document.querySelectorAll('#test-root :nth-child(odd of [attribute1], [attribute2=value2])')[14].id is "testcase29"
PASS styledElements is ["testcase1","testcase3","testcase5","testcase7","testcase9","testcase11","testcase13","testcase15","testcase17","testcase19","testcase21","testcase23","testcase25","testcase27","testcase29"]

Testing ":nth-child(even of [attribute3$="3"], :empty)"
PASS document.querySelectorAll('#test-root :nth-child(even of [attribute3$="3"], :empty)').length is 7
PASS document.querySelectorAll('#test-root :nth-child(even of [attribute3$="3"], :empty)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :nth-child(even of [attribute3$="3"], :empty)')[1].id is "testcase6"
PASS document.querySelectorAll('#test-root :nth-child(even of [attribute3$="3"], :empty)')[2].id is "testcase10"
PASS document.querySelectorAll('#test-root :nth-child(even of [attribute3$="3"], :empty)')[3].id is "testcase16"
PASS document.querySelectorAll('#test-root :nth-child(even of [attribute3$="3"], :empty)')[4].id is "testcase20"
PASS document.querySelectorAll('#test-root :nth-child(even of [attribute3$="3"], :empty)')[5].id is "testcase24"
PASS document.querySelectorAll('#test-root :nth-child(even of [attribute3$="3"], :empty)')[6].id is "testcase27"
PASS styledElements is ["testcase2","testcase6","testcase10","testcase16","testcase20","testcase24","testcase27"]

Testing ":nth-child(even of [id][attribute3$="3"], [id]:empty)"
PASS document.querySelectorAll('#test-root :nth-child(even of [id][attribute3$="3"], [id]:empty)').length is 7
PASS document.querySelectorAll('#test-root :nth-child(even of [id][attribute3$="3"], [id]:empty)')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :nth-child(even of [id][attribute3$="3"], [id]:empty)')[1].id is "testcase6"
PASS document.querySelectorAll('#test-root :nth-child(even of [id][attribute3$="3"], [id]:empty)')[2].id is "testcase10"
PASS document.querySelectorAll('#test-root :nth-child(even of [id][attribute3$="3"], [id]:empty)')[3].id is "testcase16"
PASS document.querySelectorAll('#test-root :nth-child(even of [id][attribute3$="3"], [id]:empty)')[4].id is "testcase20"
PASS document.querySelectorAll('#test-root :nth-child(even of [id][attribute3$="3"], [id]:empty)')[5].id is "testcase24"
PASS document.querySelectorAll('#test-root :nth-child(even of [id][attribute3$="3"], [id]:empty)')[6].id is "testcase27"
PASS styledElements is ["testcase2","testcase6","testcase10","testcase16","testcase20","testcase24","testcase27"]

Testing ":nth-child(even of [id][attribute2*="alue"], .foo.bar)"
PASS document.querySelectorAll('#test-root :nth-child(even of [id][attribute2*="alue"], .foo.bar)').length is 9
PASS document.querySelectorAll('#test-root :nth-child(even of [id][attribute2*="alue"], .foo.bar)')[0].id is "testcase4"
PASS document.querySelectorAll('#test-root :nth-child(even of [id][attribute2*="alue"], .foo.bar)')[1].id is "testcase6"
PASS document.querySelectorAll('#test-root :nth-child(even of [id][attribute2*="alue"], .foo.bar)')[2].id is "testcase10"
PASS document.querySelectorAll('#test-root :nth-child(even of [id][attribute2*="alue"], .foo.bar)')[3].id is "testcase14"
PASS document.querySelectorAll('#test-root :nth-child(even of [id][attribute2*="alue"], .foo.bar)')[4].id is "testcase16"
PASS document.querySelectorAll('#test-root :nth-child(even of [id][attribute2*="alue"], .foo.bar)')[5].id is "testcase20"
PASS document.querySelectorAll('#test-root :nth-child(even of [id][attribute2*="alue"], .foo.bar)')[6].id is "testcase24"
PASS document.querySelectorAll('#test-root :nth-child(even of [id][attribute2*="alue"], .foo.bar)')[7].id is "testcase27"
PASS document.querySelectorAll('#test-root :nth-child(even of [id][attribute2*="alue"], .foo.bar)')[8].id is "testcase30"
PASS styledElements is ["testcase4","testcase6","testcase10","testcase14","testcase16","testcase20","testcase24","testcase27","testcase30"]

PASS successfullyParsed is true

TEST COMPLETE
